refactor: enhance course management by centralizing configuration in AppConfig and introducing LecturerOption DTO
